Updating Faraday Gem for Compatibility with Ruby 3
This commit updates the Faraday gem for compatibility for Ruby 3.
For now I've pinned it to the latest version.

This commit also updates the author and email to match the current
owner of the Gem (with all due respect to Rune who created it).

I also delete the .travis.yaml file because we no longer use that service.